frontend-app/cloud_datastore/webapp/requirements.txt
Flask==0.12.2
google-api-python-client
